refactor: layout
-Reorganize files into a new folder structure that includes a Navigation folder and styles subfolders for each component folder.
