+++ src/components/Layout/Projects.js
@@ -9,7 +9,12 @@ function Projects() {
     <div id="projects" className={Styles.container}>
       <h2 className={Styles.title}>Projects</h2>
       {ProjData.map((data) => (
-        <div data-testid={`proj-section_${data.projId}`} id={data.projId} key={data.id} className={Styles[data.projCSS]}>
+        <div
+          data-testid={`proj-section_${data.projId}`}
+          id={data.projId}
+          key={data.id}
+          className={Styles[data.projCSS]}
+        >
           <div className={Styles.details}>
             <h3 className={Styles["project-title"]}>{data.title}</h3>
             <p className={Styles["project-description"]}>{data.description}</p>
