+++ src/components/Navigation/styles/ButtonLink.module.css
@@ -85,7 +85,6 @@
   padding: 0;
 }
 
-
 @media (max-width: 1200px) {
   .project-link {
     font-size: 1rem;
