+++ src/components/Navigation/styles/NavBar.module.css
@@ -12,6 +12,8 @@
     position: fixed;
     justify-content: center;
     align-items: center;
+    padding-top: 0.25em;
+    padding-bottom: 0.25em;
     width: 100%;
 }
