+++ src/components/Pages/Home.js
@@ -1,11 +1,10 @@
 import { React } from "react";
-import NavBar from "../UI/NavBar";
 import Skills from "../Layout/Skills";
 import Projects from "../Layout/Projects";
 import Contact from "../Layout/Contact";
 import Footer from "../Layout/Footer";
 import Styles from "./Home.module.css";
-import Button from "../UI/LinkButton";
+import Button from "../Navigation/ButtonLink";
 import HomeBtnData from "../../data/HomeBtnData";
 import ProfilePic from "../../assets/profile-pic.png";
 import Modal from "../UI/Modal";
@@ -33,24 +32,17 @@ function Home({ handleModal, showModal }) {
             <span onClick={handleModal} className={Styles["text-link"]}>
               Electrical Engineering and Management
             </span>{" "}
-            from Ontario Tech University, who loves collaborating, creating,
-            learning, and problem-solving. For instance, during my capstone
-            project, I managed a team of three students while programming an
-            Arduino Arc Fault detection system and mobile application, enabling
-            real-time notifications through an SQL database.
+            who loves creating, collaborating, and problem-solving.
           </p>
           <p className={Styles.description}>
-            I am dedicated to continuous learning and have obtained
+            I'm dedicated to continuous learning and have obtained
             certifications from FreeCodeCamp (Front End Developer Certification)
             and Harvard University (CS50 Introduction to Computer Science).
-            Furthermore, I have undertaken a multitude of programming courses
-            and projects, significantly expanding my expertise across various
-            facets of software development.
           </p>
           <p className={Styles.description}>
-            In addition to my passion for programming, I have a variety of
-            interests including archery, badminton, drawing, gaming, exercise,
-            and getting quality sleep.
+            If I'm not programming, I'm playing video games, exercising, or
+            spend time with friends and family on fun activities like archery,
+            badminton, chess, and board games.
           </p>
           <ul className={Styles.links}>
             {HomeBtnData.map((data, index) => (
