+++ src/components/UI/Modal.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Styles from "./Modal.module.css";
 import uniLogoImg from "../../assets/ontariotechu-img.jpg";
-import Button from "./LinkButton";
+import Button from "../Navigation/ButtonLink";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faXmark as closeIcon } from "@fortawesome/free-solid-svg-icons";
